Refactor testing infrastructure with unified test runner and workflow
