--- tests/test_config.py
@@ -1,303 +1,0 @@
-#!/usr/bin/env python3
-"""
-Unit tests for configuration functionality
-"""
-
-import unittest
-import tempfile
-import os
-import json
-import shutil
-from pathlib import Path
-from c_to_plantuml.config import Config
-from c_to_plantuml.models import ProjectModel, FileModel, Struct, Enum, Function, Field
-
-
-class TestConfig(unittest.TestCase):
-    """Test cases for configuration functionality"""
-    
-    def setUp(self):
-        """Set up test fixtures"""
-        self.temp_dir = tempfile.mkdtemp()
-        self.test_files = []
-    
-    def tearDown(self):
-        """Clean up test fixtures"""
-        shutil.rmtree(self.temp_dir, ignore_errors=True)
-    
-    def create_test_config(self, config_data: dict) -> str:
-        """Create a test configuration file and return its path"""
-        config_path = os.path.join(self.temp_dir, "test_config.json")
-        with open(config_path, 'w', encoding='utf-8') as f:
-            json.dump(config_data, f, indent=2)
-        return config_path
-    
-    def test_load_valid_config(self):
-        """Test loading a valid configuration file"""
-        config_data = {
-            "project_name": "test_project",
-            "project_roots": ["/path/to/project"],
-            "output_dir": "./output",
-            "model_output_path": "model.json",
-            "recursive": True
-        }
-        
-        config_path = self.create_test_config(config_data)
-        config = Config.load(config_path)
-        
-        self.assertEqual(config.project_name, "test_project")
-        self.assertEqual(config.project_roots, ["/path/to/project"])
-        self.assertEqual(config.output_dir, "./output")
-        self.assertEqual(config.model_output_path, "model.json")
-        self.assertTrue(config.recursive)
-    
-    def test_load_config_with_filters(self):
-        """Test loading configuration with file and element filters"""
-        config_data = {
-            "project_name": "test_project",
-            "project_roots": ["/path/to/project"],
-            "file_filters": {
-                "include": [".*\\.c$", ".*\\.h$"],
-                "exclude": ["test_.*\\.c$"]
-            },
-            "element_filters": {
-                "structs": {
-                    "include": ["Person", "Config"],
-                    "exclude": ["Internal.*"]
-                }
-            }
-        }
-        
-        config_path = self.create_test_config(config_data)
-        config = Config.load(config_path)
-        
-        self.assertTrue(config.has_filters())
-        self.assertEqual(len(config.file_include_patterns), 2)
-        self.assertEqual(len(config.file_exclude_patterns), 1)
-    
-    def test_config_validation(self):
-        """Test configuration validation"""
-        # Test invalid project_roots (not a list)
-        config_data = {
-            "project_name": "test_project",
-            "project_roots": "invalid_string"
-        }
-        
-        config_path = self.create_test_config(config_data)
-        with self.assertRaises(ValueError):
-            Config.load(config_path)
-    
-    def test_save_and_load_config(self):
-        """Test saving and loading configuration"""
-        original_config = Config({
-            "project_name": "test_project",
-            "project_roots": ["/path/to/project"],
-            "output_dir": "./output",
-            "recursive": False,
-            "file_filters": {
-                "include": [".*\\.c$"]
-            }
-        })
-        
-        # Save configuration
-        config_path = os.path.join(self.temp_dir, "saved_config.json")
-        original_config.save(config_path)
-        
-        # Load configuration
-        loaded_config = Config.load(config_path)
-        
-        # Verify loaded configuration matches original
-        self.assertEqual(loaded_config.project_name, original_config.project_name)
-        self.assertEqual(loaded_config.project_roots, original_config.project_roots)
-        self.assertEqual(loaded_config.output_dir, original_config.output_dir)
-        self.assertEqual(loaded_config.recursive, original_config.recursive)
-        self.assertEqual(len(loaded_config.file_include_patterns), 1)
-    
-    def test_file_filtering(self):
-        """Test file filtering functionality"""
-        config = Config({
-            "file_filters": {
-                "include": [".*\\.c$", ".*\\.h$"],
-                "exclude": ["test_.*\\.c$", ".*\\.tmp$"]
-            }
-        })
-        
-        # Test include patterns
-        self.assertTrue(config._should_include_file("main.c"))
-        self.assertTrue(config._should_include_file("header.h"))
-        self.assertFalse(config._should_include_file("main.cpp"))
-        
-        # Test exclude patterns
-        self.assertFalse(config._should_include_file("test_main.c"))
-        self.assertFalse(config._should_include_file("file.tmp"))
-    
-    def test_element_filtering(self):
-        """Test element filtering functionality"""
-        config = Config({
-            "element_filters": {
-                "structs": {
-                    "include": ["Person", "Config"],
-                    "exclude": ["Internal.*"]
-                },
-                "functions": {
-                    "include": ["main", "process"],
-                    "exclude": ["debug_.*"]
-                }
-            }
-        })
-        
-        # Create test file model
-        file_model = FileModel(
-            file_path="test.c",
-            relative_path="test.c",
-            project_root="/test",
-            encoding_used="utf-8",
-            structs={
-                "Person": Struct("Person", []),
-                "Config": Struct("Config", []),
-                "InternalData": Struct("InternalData", [])
-            },
-            enums={},
-            functions=[
-                Function("main", "int", []),
-                Function("process", "void", []),
-                Function("debug_log", "void", [])
-            ],
-            globals=[],
-            includes=[],
-            macros=[],
-            typedefs={}
-        )
-        
-        # Apply filters
-        filtered_model = config._apply_element_filters(file_model)
-        
-        # Check struct filtering
-        self.assertIn("Person", filtered_model.structs)
-        self.assertIn("Config", filtered_model.structs)
-        self.assertNotIn("InternalData", filtered_model.structs)
-        
-        # Check function filtering
-        function_names = [f.name for f in filtered_model.functions]
-        self.assertIn("main", function_names)
-        self.assertIn("process", function_names)
-        self.assertNotIn("debug_log", function_names)
-    
-    def test_model_filtering(self):
-        """Test applying filters to a complete model"""
-        config = Config({
-            "file_filters": {
-                "include": [".*\\.c$"],
-                "exclude": ["test_.*\\.c$"]
-            },
-            "element_filters": {
-                "structs": {
-                    "include": ["Person"]
-                }
-            }
-        })
-        
-        # Create test model
-        file1 = FileModel(
-            file_path="main.c",
-            relative_path="main.c",
-            project_root="/test",
-            encoding_used="utf-8",
-            structs={
-                "Person": Struct("Person", []),
-                "Config": Struct("Config", [])
-            },
-            enums={},
-            functions=[],
-            globals=[],
-            includes=[],
-            macros=[],
-            typedefs={}
-        )
-        
-        file2 = FileModel(
-            file_path="test_helper.c",
-            relative_path="test_helper.c",
-            project_root="/test",
-            encoding_used="utf-8",
-            structs={},
-            enums={},
-            functions=[],
-            globals=[],
-            includes=[],
-            macros=[],
-            typedefs={}
-        )
-        
-        model = ProjectModel(
-            project_name="test",
-            project_root="/test",
-            files={
-                "main.c": file1,
-                "test_helper.c": file2
-            },
-            created_at="2024-01-01T00:00:00"
-        )
-        
-        # Apply filters
-        filtered_model = config.apply_filters(model)
-        
-        # Check file filtering
-        self.assertIn("main.c", filtered_model.files)
-        self.assertNotIn("test_helper.c", filtered_model.files)
-        
-        # Check element filtering
-        main_file = filtered_model.files["main.c"]
-        self.assertIn("Person", main_file.structs)
-        self.assertNotIn("Config", main_file.structs)
-    
-    def test_invalid_regex_patterns(self):
-        """Test handling of invalid regex patterns"""
-        config_data = {
-            "project_name": "test_project",
-            "project_roots": ["/path/to/project"],
-            "file_filters": {
-                "include": ["[invalid_regex"],
-                "exclude": ["valid_pattern"]
-            }
-        }
-        
-        config_path = self.create_test_config(config_data)
-        # Should not raise an exception, but log a warning
-        config = Config.load(config_path)
-        
-        # Should have compiled the valid pattern but not the invalid one
-        self.assertEqual(len(config.file_include_patterns), 0)
-        self.assertEqual(len(config.file_exclude_patterns), 1)
-    
-    def test_get_summary(self):
-        """Test configuration summary generation"""
-        config = Config({
-            "project_name": "test_project",
-            "project_roots": ["/path1", "/path2"],
-            "output_dir": "./output",
-            "recursive": True,
-            "file_filters": {
-                "include": [".*\\.c$"]
-            },
-            "element_filters": {
-                "structs": {
-                    "include": ["Person"]
-                }
-            }
-        })
-        
-        summary = config.get_summary()
-        
-        self.assertEqual(summary["project_name"], "test_project")
-        self.assertEqual(summary["project_roots"], ["/path1", "/path2"])
-        self.assertEqual(summary["output_dir"], "./output")
-        self.assertTrue(summary["recursive"])
-        self.assertTrue(summary["has_file_filters"])
-        self.assertTrue(summary["has_element_filters"])
-        self.assertEqual(summary["include_patterns"], 1)
-        self.assertEqual(summary["exclude_patterns"], 0)
-
-
-if __name__ == '__main__':
-    unittest.main()
